Award Timeline.
Added timeline under Award (under Portfolio) tab. Information are
incomplete and will be updated once information is given. The icons as
well will be edited if required. For the Contacts Section, animation for
icon Address and icon Operating Hours are removed as there are no link
for those icons. Links are added to the rest of the Contact icons.